Set 'debug' as log level for testing
Loggy is no longer handling messages for levels below the configured
ones, hence we're getting failing tests because 'info' messages are not
showing up (correctly) since the log level used to be 'error'. This
commit changes the minimum level to 'debug' so that we can test anything
we want around logging, but also hides the output to avoid very noisy
tests.