Paginate through results
This adds support for large result set to FileStorage, however the
pagination is happening behind the scenes and not exposed to the caller
which will only see the final result set via an enumerable).

This is a breaking change as previously the full result set was a
returned in a list.