Refactor FileStorage main interface
This commit moves the main `#for` method of `FileStorage` inside the top
level `FileStorage` module and moves the implementation of `KeyCtx` (now
renamed to `KeyContext`) and `KeyStorage` in their own classes.

There's no functional changes as this is just a small reorganisation of
the code layout, but should allow better testing of the individual
classes.