Support escapable URIs
Passing an URI that needs to be escaped to BucketStore will result in an
error. For example the following will raise an `URI::InvalidURIError`.

```ruby
BucketStore.for("s3://bucket/a key with spaces.txt")
```

This commit adds support for internal escape/unescape of URIs, meaning
that external callers don't need to worry about passing unescaped values
to `.for`.

Fixes #46