Allow File uploads instead of only supporting Strings
To support uploading of large files this change allows an [IO](https://ruby-doc.org/core-2.7.1/IO.html) object to be provided instead of requiring a string. This allows large files to be streamed into the relevant backing store. The S3 / GCS SDKs already support this as a first class concept.

Prior to this change the individual adapters `upload!` method only supported the notion of uploading a string. With this change we that method think in terms of IO like objects with String uploads being a special case where we need to wrap the String in a `StringIO`.

The interface is kept backwards compatible through exposing a separate `upload_file!` method alongside the existing `upload!` method.