Add KeyStreamer interface for uploads and downloads.
Completely stolen from https://github.com/gocardless/bucket-store/pull/63.

The KeyStreamer expects uploads / downloads to be expressed in terms of IO operations.

For example a download is actually `download into this IO` and upload is `upload the content from this file`.

Uploads / downloads of Strings are considered a special case of this. However is distinct enough that we have streaming be it's own API.