Lazy-load client libraries.
Currently when loading the gem we also implicitly load the client
libraries for the bucket providers. These are quite large, and increase
the time to load the gem (and the number of files included) by orders of
magnitude.

To avoid unnecessary slowing down the bootstrap, we can instead load
these gems only if the provider is used, and memoise this on the class
to avoid repeated calls to `require` (as this returns `true` when the
library is first required).

In testing with Rails, lazy-loading these libraries reduces the time
taken for Bundler to require the gem during bootstrap from 0.25 to 0.005
seconds (yes, that's two orders of magnitude faster).